Add request-level escrow and redundancy-aware payouts
- Per-request locks in Wallet (lockForRequest / disburseForRequest / releaseForRequest) for precise fund reservation
- SubscriptionsManager now uses request-level lock/release semantics and delegates commitment issuance to the coordinator

+++ src/v1_0_0/consumer/BaseConsumer.sol
@@ -50,7 +50,7 @@ abstract contract BaseConsumer is Routable {
     /// @param containerId if lazy subscription, subscription compute container ID, else empty
     /// @param index if lazy subscription, `Inbox` lazy store index, else empty
     function _receiveCompute(
-        uint32 subscriptionId,
+        uint64 subscriptionId,
         uint32 interval,
         uint16 redundancy,
         address node,
@@ -68,7 +68,7 @@ abstract contract BaseConsumer is Routable {
     /// @param timestamp timestamp at which container inputs are collected
     /// @param caller calling address
     /// @return The container inputs as bytes.
-    function getContainerInputs(uint32 subscriptionId, uint32 interval, uint32 timestamp, address caller)
+    function getContainerInputs(uint64 subscriptionId, uint32 interval, uint32 timestamp, address caller)
         external
         view
         virtual
